feat: 登录，新增员工

+++ take-out-pojo/src/main/java/com/player32611/entity/Category.java
@@ -10,13 +10,13 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Category {
-    private Integer id;
+    private Long id;
     private String name;
     private Integer type;
     private Integer sort;
     private Integer status;
     private LocalDateTime create_time;
     private LocalDateTime update_time;
-    private Integer create_user;
-    private Integer update_user;
+    private Long create_user;
+    private Long update_user;
 }
