+++ take-out-pojo/src/main/java/com/player32611/entity/Dish.java
@@ -11,15 +11,15 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Dish {
-    private Integer id;
+    private Long id;
     private String name;
-    private Integer category_id;
+    private Long category_id;
     private BigDecimal price;
     private String image;
     private String description;
     private Integer status;
     private LocalDateTime create_time;
     private LocalDateTime update_time;
-    private Integer create_user;
-    private Integer update_user;
+    private Long create_user;
+    private Long update_user;
 }
